Decouple ParameterSet and lua
