+++ core/include/lua/ParameterSetParser.h
@@ -1,0 +1,35 @@
+/*
+ *  MoMEMta: a modular implementation of the Matrix Element Method
+ *  Copyright (C) 2016  Universite catholique de Louvain (UCL), Belgium
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#pragma once
+
+#include <momemta/ParameterSet.h>
+
+struct lua_State;
+
+class ParameterSetParser {
+public:
+
+    /**
+     * \brief Convert a lua table to a ParameterSet
+     *
+     * \param L the lua state
+     * \param index The index on the stack of the table to parse
+     */
+    static void parse(ParameterSet& p, lua_State* L, int index);
+};
